rework prefill data (#816)
